+++ build_package.sh
@@ -6,7 +6,7 @@
 set -e
 
 PLUGIN_NAME="mem0ai-local"
-VERSION="0.1.13"
+VERSION="0.2.1"
 OUTPUT_FILE="${PLUGIN_NAME}-${VERSION}.difypkg"
 TEMP_DIR="temp_package"